test: interest repositoy delete 테스트 추가
